Coop fini. Manque les avatars

+++ src/components/Membre.vue
@@ -3,11 +3,13 @@
         <div class="row">
         <div class="vr"></div>
             <div class="colmun">
-                {{membre.fullname}}
-                <a :href="'mailto'+membre.email">{{membre.email}}</a>
+                <p>{{membre.fullname}}</p>
+                <a :href="'mailto:'+membre.email">{{membre.email}}</a>
             </div>
             <div class="column">
-                <button class="button" title="Profil">👤</button>
+                <router-link :to="{name: 'Profil', params: { membre_id: membre.id },}">
+                    <button class="button" title="Profil">👤</button>
+                </router-link>
                 <button class="button button-outline" @click="effacerMembre" title="Supprimer">🗑</button>
             </div>
             <div class="vr"></div>
@@ -25,14 +27,19 @@ export default {
     },
     methods : {
         effacerMembre(){
-            if(confirm('Voulez-vous supprimer ce membre : '+this.membre.fullname+' ? ')){
-                api.delete('members/'+this.membre.id).then(response =>{
-                    this.$bus.$emit('charger-membres');
-                }).catch(error => {
-                    console.log(error.response.data);
-                })
+            if(this.$store.state.membre.id != this.membre.id){
+                if(confirm('Voulez-vous supprimer ce membre : '+this.membre.fullname+' ? ')){
+                    api.delete('members/'+this.membre.id).then(response =>{
+                        this.$bus.$emit('charger-membres');
+                    }).catch(error => {
+                        console.log(error.response.data);
+                    })
+                }
+            } else {
+                alert("Impossible de vous supprimer vous-même.");
             }
-        } 
+            
+        }
     }
 }
 </script>
